separate components into dirs
