+++ example/app/app.component.html
@@ -5,7 +5,7 @@ I was really hoping for a component api like:
 ```html
 <supre-card-navbar>
   <supre-card-navbar-item>
-    <supre-card-navbar-title>Segments</supre-card-navbar-top-menu-item>
+    <supre-card-navbar-title>Segments</supre-card-navbar-menu-item>
     <supre-card-navbar-cards>
       <supre-card-navbar-card>
         <supre-card-navbar-card-title>Library</supre-card-navbar-card-title>
@@ -17,7 +17,7 @@ I was really hoping for a component api like:
 </supre-card-navbar>
 ```
 
-And then have the component split the titles into a `navbar-top-items` div,
+And then have the component split the titles into a `navbar-items` div,
 and the cards into a `navbar-sub-items` div. This would allow the item
 and cards to be coupled in the api, but distinct in the markup (out of necessity)
 
@@ -25,7 +25,7 @@ Unfortanely, I was not able to accomplish this with ng2.
 -->
 
 <supre-card-navbar>
-  <supre-card-navbar-top-menu-item supreTabId="segments" [supreDefault]="true">Segments</supre-card-navbar-top-menu-item>
+  <supre-card-navbar-menu-item supreTabId="segments" [supreDefault]="true">Segments</supre-card-navbar-menu-item>
   <supre-card-navbar-cards supreForTab="segments">
     <supre-card-navbar-card supreForTab="segments" [supreDefault]="true" [supreUltimateDefault]="true">
       <supre-card-navbar-card-title>Create</supre-card-navbar-card-title>
@@ -44,7 +44,7 @@ Unfortanely, I was not able to accomplish this with ng2.
       <supre-card-navbar-card-icon><div class="icon ion-ios-glasses-outline"></div></supre-card-navbar-card-icon>
     </supre-card-navbar-card>
   </supre-card-navbar-cards>
-  <supre-card-navbar-top-menu-item supreTabId="campaigns">Campaigns</supre-card-navbar-top-menu-item>
+  <supre-card-navbar-menu-item supreTabId="campaigns">Campaigns</supre-card-navbar-menu-item>
   <supre-card-navbar-cards supreForTab="campaigns">
     <supre-card-navbar-card supreForTab="campaigns" [supreDefault]="true">
       <supre-card-navbar-card-title>Option</supre-card-navbar-card-title>
@@ -59,14 +59,14 @@ Unfortanely, I was not able to accomplish this with ng2.
       <supre-card-navbar-card-icon><div class="icon ion-record"></div></supre-card-navbar-card-icon>
     </supre-card-navbar-card>
   </supre-card-navbar-cards>
-  <supre-card-navbar-top-menu-item supreTabId="inventory">Inventory</supre-card-navbar-top-menu-item>
+  <supre-card-navbar-menu-item supreTabId="inventory">Inventory</supre-card-navbar-menu-item>
   <supre-card-navbar-cards supreForTab="inventory">
     <supre-card-navbar-card supreForTab="inventory" [supreDefault]="true">
       <supre-card-navbar-card-title>Choice</supre-card-navbar-card-title>
       <supre-card-navbar-card-icon><div class="icon ion-record"></div></supre-card-navbar-card-icon>
     </supre-card-navbar-card>
   </supre-card-navbar-cards>
-  <supre-card-navbar-top-menu-item supreTabId="user">Mary Popins</supre-card-navbar-top-menu-item>
+  <supre-card-navbar-menu-item supreTabId="user">Mary Popins</supre-card-navbar-menu-item>
   <supre-card-navbar-cards supreForTab="user">
     <supre-card-navbar-card supreForTab="user">
       <supre-card-navbar-card-title>Log Out</supre-card-navbar-card-title>
