--- src/card-navbar-card-icon.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { Subject }   from 'rxjs/Subject';
-
-@Component({
-  selector: 'supre-card-navbar-card-icon',
-  template: require('./card-navbar-card-icon.component.html'),
-  styles: [require('./card-navbar-card-icon.component.css')]
-})
-export class CardNavbarCardIconComponent {
-
-}
