--- src/card-navbar-card-title.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { Subject }   from 'rxjs/Subject';
-
-@Component({
-  selector: 'supre-card-navbar-card-title',
-  template: require('./card-navbar-card-title.component.html'),
-  styles: [require('./card-navbar-card-title.component.css')]
-})
-export class CardNavbarCardTitleComponent {
-
-}
