--- src/card-navbar-menu-cards.component.scss
@@ -1,3 +1,0 @@
-$blue: rgba(0, 137, 208, 1);
-$blue-transparent: rgba(0, 137, 208, .3);
-$gray: #999999;
