--- src/card-navbar-menu-cards.component.ts
@@ -1,1 +1,0 @@
-// delete file
