--- src/card-navbar-user-menu-item.component.css
@@ -1,64 +1,0 @@
-:host {
-  height: 100%; }
-
-:host[supretabid="user"] {
-  margin-left: auto; }
-
-.CardNavbar-topMenuItem {
-  cursor: pointer;
-  display: flex;
-  justify-content: center; }
-
-.CardNavbar-topMenuItemContainer {
-  margin-left: 90px;
-  height: 100%; }
-
-.CardNavbar-topMenuItem {
-  font-weight: 600;
-  font-size: 10px;
-  font-family: Verdana;
-  display: inline-block;
-  text-transform: uppercase;
-  color: white;
-  height: 100%;
-  min-width: 100px;
-  display: flex;
-  align-items: center; }
-
-.CardNavbar-topMenuItem.is-selected {
-  color: #0089d0;
-  background-color: #e6e6e6; }
-
-.CardNavbar-topMenuItem.is-active {
-  color: #0089d0;
-  background-color: white; }
-
-.CardNavbar-userMenuItem {
-  font-size: 11px;
-  width: auto;
-  justify-content: flex-end; }
-
-.CardNavbar-userMenuItem.is-selected,
-.CardNavbar-userMenuItem.is-active {
-  width: 300px; }
-
-.CardNavbar-userMenuIcon {
-  margin-right: 20px;
-  margin-left: 10px;
-  font-size: 26px; }
-
-@media (max-width: 980px) {
-  .CardNavbar-topMenuItemContainer {
-    margin-left: 50px; } }
-
-:host {
-  margin-left: auto; }
-
-.CardNavbar-userMenuItem {
-  font-size: 11px;
-  width: auto; }
-
-.CardNavbar-userMenuIcon {
-  margin-right: 20px;
-  margin-left: 10px;
-  font-size: 26px; }
