--- src/card-navbar-user-menu-item.component.html
@@ -1,4 +1,0 @@
-<div class="CardNavbar-topMenuItem CardNavbar-userMenuItem">
-  <div class="CardNavbar-userMenuName"><ng-content></ng-content></div>
-  <div class="CardNavbar-userMenuIcon icon ion-android-person"></div>
-</div>
