--- src/card-navbar-user-menu-item.component.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-import { Subject }   from 'rxjs/Subject';
-
-@Component({
-  selector: 'supre-card-navbar-user-menu-item',
-  template: require('./card-navbar-user-menu-item.component.html'),
-  styles: [require('./card-navbar-user-menu-item.component.css')]
-})
-export class CardNavbarUserMenuItemComponent {
-
-}
