+++ src/card-navbar/card-navbar.component.html
@@ -1,11 +1,11 @@
 <div class="GiraffeLogo">
-  <img src="images/the-G-raph.svg">
+  <img src="../images/the-G-raph.svg">
 </div>
 
 <div class="CardNavbar">
   <div class="CardNavbar-navbar">
-    <img class="CardNavbar-title" src="images/g-raph_type.svg">
-    <ng-content select="supre-card-navbar-top-menu-item"></ng-content>
+    <img class="CardNavbar-title" src="../images/g-raph_type.svg">
+    <ng-content select="supre-card-navbar-menu-item"></ng-content>
   </div>
   <div class="CardNavbar-cards js-cards">
     <ng-content select="supre-card-navbar-cards"></ng-content>
