+++ src/card-navbar/card-navbar.component.scss
