+++ src/card-navbar/cards/card/card-icon/card-icon.component.scss
