+++ src/card-navbar/cards/card/card-icon/card-icon.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'supre-card-navbar-card-icon',
+  template: require('./card-icon.component.html'),
+  styles: [require('./card-icon.component.css')]
+})
+export class CardNavbarCardIconComponent {}
