+++ src/card-navbar/cards/card/card-title/card-title.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'supre-card-navbar-card-title',
+  template: require('./card-title.component.html'),
+  styles: [require('./card-title.component.css')]
+})
+export class CardNavbarCardTitleComponent {}
