+++ src/card-navbar/cards/card/card.component.scss
