+++ src/card-navbar/menu-item/menu-item.component.scss
