+++ src/index.ts
@@ -1,22 +1,20 @@
 import { CommonModule } from '@angular/common';
 import { StateManagerService } from './state-manager.service';
-import { CardNavbarComponent } from './card-navbar.component';
-import { CardNavbarTopMenuItemComponent } from './card-navbar-top-menu-item.component';
-import { CardNavbarUserMenuItemComponent } from './card-navbar-user-menu-item.component';
-import { CardNavbarCardsComponent } from './card-navbar-cards.component';
-import { CardNavbarCardComponent } from './card-navbar-card.component';
-import { CardNavbarCardTitleComponent } from './card-navbar-card-title.component';
-import { CardNavbarCardIconComponent } from './card-navbar-card-icon.component';
+import { CardNavbarComponent } from './card-navbar/card-navbar.component';
+import { CardNavbarMenuItemComponent } from './card-navbar/menu-item/menu-item.component';
+import { CardNavbarCardsComponent } from './card-navbar/cards/cards.component';
+import { CardNavbarCardComponent } from './card-navbar/cards/card/card.component';
+import { CardNavbarCardTitleComponent } from './card-navbar/cards/card/card-title/card-title.component';
+import { CardNavbarCardIconComponent } from './card-navbar/cards/card/card-icon/card-icon.component';
 import { NgModule } from '@angular/core';
 
 export * from './state-manager.service';
-export * from './card-navbar.component';
-export * from './card-navbar-top-menu-item.component';
-export * from './card-navbar-user-menu-item.component';
-export * from './card-navbar-cards.component';
-export * from './card-navbar-card.component';
-export * from './card-navbar-card-title.component';
-export * from './card-navbar-card-icon.component';
+export * from './card-navbar/card-navbar.component';
+export * from './card-navbar/menu-item/menu-item.component';
+export * from './card-navbar/cards/cards.component';
+export * from './card-navbar/cards/card/card.component';
+export * from './card-navbar/cards/card/card-title/card-title.component';
+export * from './card-navbar/cards/card/card-icon/card-icon.component';
 
 @NgModule({
   imports: [
@@ -27,8 +25,7 @@ export * from './card-navbar-card-icon.component';
   ],
   declarations: [
     CardNavbarComponent,
-    CardNavbarTopMenuItemComponent,
-    CardNavbarUserMenuItemComponent,
+    CardNavbarMenuItemComponent,
     CardNavbarCardsComponent,
     CardNavbarCardComponent,
     CardNavbarCardTitleComponent,
@@ -36,8 +33,7 @@ export * from './card-navbar-card-icon.component';
   ],
   exports: [
     CardNavbarComponent,
-    CardNavbarTopMenuItemComponent,
-    CardNavbarUserMenuItemComponent,
+    CardNavbarMenuItemComponent,
     CardNavbarCardsComponent,
     CardNavbarCardComponent,
     CardNavbarCardTitleComponent,
@@ -45,8 +41,7 @@ export * from './card-navbar-card-icon.component';
   ],
   entryComponents: [
     CardNavbarComponent,
-    CardNavbarTopMenuItemComponent,
-    CardNavbarUserMenuItemComponent,
+    CardNavbarMenuItemComponent,
     CardNavbarCardsComponent,
     CardNavbarCardComponent,
     CardNavbarCardTitleComponent,
